[#3745] Reduce double code between ChangeDependencyGroupIdAndArtifactId, ChangeManagedDependencyGroupIdAndArtifactId and ChangePluginGroupIdAndArtifactId
