+++ rewrite-maven/src/main/java/org/openrewrite/maven/ChangeDependencyGroupIdAndArtifactId.java
@@ -16,30 +16,19 @@
 package org.openrewrite.maven;
 
 import static org.openrewrite.Validated.required;
-import static org.openrewrite.Validated.test;
-import static org.openrewrite.internal.StringUtils.isBlank;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
 import java.util.NoSuchElementException;
-import java.util.Objects;
 import java.util.Optional;
 
+import org.jetbrains.annotations.NotNull;
 import org.openrewrite.ExecutionContext;
 import org.openrewrite.Option;
-import org.openrewrite.Recipe;
 import org.openrewrite.TreeVisitor;
 import org.openrewrite.Validated;
 import org.openrewrite.internal.lang.Nullable;
-import org.openrewrite.maven.table.MavenMetadataFailures;
-import org.openrewrite.maven.tree.MavenMetadata;
 import org.openrewrite.maven.tree.ResolvedManagedDependency;
 import org.openrewrite.maven.tree.Scope;
-import org.openrewrite.semver.Semver;
-import org.openrewrite.semver.VersionComparator;
 import org.openrewrite.xml.AddToTagVisitor;
-import org.openrewrite.xml.ChangeTagValueVisitor;
 import org.openrewrite.xml.RemoveContentVisitor;
 import org.openrewrite.xml.tree.Xml;
 
@@ -50,48 +39,7 @@ import lombok.Value;
 
 @Value
 @EqualsAndHashCode(callSuper = false)
-public class ChangeDependencyGroupIdAndArtifactId extends Recipe {
-    transient MavenMetadataFailures metadataFailures = new MavenMetadataFailures(this);
-
-    @Option(displayName = "Old groupId",
-            description = "The old groupId to replace. The groupId is the first part of a dependency coordinate `com.google.guava:guava:VERSION`. Supports glob expressions.",
-            example = "org.openrewrite.recipe")
-    String oldGroupId;
-
-    @Option(displayName = "Old artifactId",
-            description = "The old artifactId to replace. The artifactId is the second part of a dependency coordinate `com.google.guava:guava:VERSION`. Supports glob expressions.",
-            example = "rewrite-testing-frameworks")
-    String oldArtifactId;
-
-    @Option(displayName = "New groupId",
-            description = "The new groupId to use. Defaults to the existing group id.",
-            example = "corp.internal.openrewrite.recipe",
-            required = false)
-    @Nullable
-    String newGroupId;
-
-    @Option(displayName = "New artifactId",
-            description = "The new artifactId to use. Defaults to the existing artifact id.",
-            example = "rewrite-testing-frameworks",
-            required = false)
-    @Nullable
-    String newArtifactId;
-
-    @Option(displayName = "New version",
-            description = "An exact version number or node-style semver selector used to select the version number.",
-            example = "29.X",
-            required = false)
-    @Nullable
-    String newVersion;
-
-    @Option(displayName = "Version pattern",
-            description = "Allows version selection to be extended beyond the original Node Semver semantics. So for example," +
-                          "Setting 'version' to \"25-29\" can be paired with a metadata pattern of \"-jre\" to select Guava 29.0-jre",
-            example = "-jre",
-            required = false)
-    @Nullable
-    String versionPattern;
-
+public class ChangeDependencyGroupIdAndArtifactId extends AbstractChangeDependencyGroupIdAndArtifactId {
     @Option(displayName = "Override managed version",
             description = "If the new dependency has a managed version, this flag can be used to explicitly set the version on the dependency. The default for this flag is `false`.",
             required = false)
@@ -104,18 +52,16 @@ public class ChangeDependencyGroupIdAndArtifactId extends Recipe {
     @Nullable
     Boolean changeManagedDependency;
 
-    public ChangeDependencyGroupIdAndArtifactId(String oldGroupId, String oldArtifactId, @Nullable String newGroupId, @Nullable String newArtifactId, @Nullable String newVersion, @Nullable String versionPattern) {
+    public ChangeDependencyGroupIdAndArtifactId(String oldGroupId, String oldArtifactId, String newGroupId, String newArtifactId,
+            @Nullable String newVersion, @Nullable String versionPattern) {
         this(oldGroupId, oldArtifactId, newGroupId, newArtifactId, newVersion, versionPattern, false, true);
     }
 
     @JsonCreator
-    public ChangeDependencyGroupIdAndArtifactId(String oldGroupId, String oldArtifactId, @Nullable String newGroupId, @Nullable String newArtifactId, @Nullable String newVersion, @Nullable String versionPattern, @Nullable Boolean overrideManagedVersion, @Nullable Boolean changeManagedDependency) {
-        this.oldGroupId = oldGroupId;
-        this.oldArtifactId = oldArtifactId;
-        this.newGroupId = newGroupId;
-        this.newArtifactId = newArtifactId;
-        this.newVersion = newVersion;
-        this.versionPattern = versionPattern;
+    public ChangeDependencyGroupIdAndArtifactId(String oldGroupId, String oldArtifactId, String newGroupId, String newArtifactId,
+            @Nullable String newVersion, @Nullable String versionPattern, @Nullable Boolean overrideManagedVersion,
+            @Nullable Boolean changeManagedDependency) {
+        super(oldGroupId, oldArtifactId, newGroupId, newArtifactId, newVersion, versionPattern);
         this.overrideManagedVersion = overrideManagedVersion;
         this.changeManagedDependency = changeManagedDependency;
     }
@@ -136,33 +82,13 @@ public class ChangeDependencyGroupIdAndArtifactId extends Recipe {
     }
 
     @Override
-    public Validated<Object> validate() {
-        Validated<Object> validated = super.validate();
-        if (newVersion != null) {
-            validated = validated.and(Semver.validate(newVersion, versionPattern));
-        }
-        validated = validated.and(required("newGroupId", newGroupId).or(required("newArtifactId", newArtifactId)));
-        validated = validated.and(test(
-                "coordinates",
-                "newGroupId OR newArtifactId must be different from before",
-                this,
-                r -> {
-                    boolean sameGroupId = isBlank(r.newGroupId) || Objects.equals(r.oldGroupId, r.newGroupId);
-                    boolean sameArtifactId = isBlank(r.newArtifactId) || Objects.equals(r.oldArtifactId, r.newArtifactId);
-                    return !(sameGroupId && sameArtifactId);
-                }
-        ));
-        return validated;
+    protected @NotNull Validated<Object> addExtraValidation(Validated<Object> validated) {
+        return validated.and(required("newGroupId", newGroupId).or(required("newArtifactId", newArtifactId)));
     }
 
     @Override
     public TreeVisitor<?, ExecutionContext> getVisitor() {
-        return new MavenVisitor<ExecutionContext>() {
-            @Nullable
-            final VersionComparator versionComparator = newVersion != null ? Semver.validate(newVersion, versionPattern).getValue() : null;
-            @Nullable
-            private Collection<String> availableVersions;
-
+        return new AbstractChangeDependencyGroupIdAndArtifactIdVisitor() {
             @Override
             public Xml visitDocument(Xml.Document document, ExecutionContext ctx) {
                 // Any managed dependency change is unlikely to use the same version, so only update selectively.
@@ -232,38 +158,12 @@ public class ChangeDependencyGroupIdAndArtifactId extends Recipe {
                 return t;
             }
 
-            private Xml.Tag changeChildTagValue(Xml.Tag tag, String childTagName, String newValue, ExecutionContext ctx) {
-                Optional<Xml.Tag> childTag = tag.getChild(childTagName);
-                if (childTag.isPresent() && !newValue.equals(childTag.get().getValue().orElse(null))) {
-                    tag = (Xml.Tag) new ChangeTagValueVisitor<>(childTag.get(), newValue).visitNonNull(tag, ctx);
-                }
-                return tag;
-            }
-
             private boolean isDependencyManaged(Scope scope, String groupId, String artifactId) {
 
 				@Nullable
                 ResolvedManagedDependency dependency = getResolutionResult().getPom().getResolvedManagedDependencyWithMinimumProximity(dm ->
                         groupId.equals(dm.getGroupId()) && artifactId.equals(dm.getArtifactId()));
                 return dependency != null && scope.isInClasspathOf(dependency.getScope());
-            }
-
-            @SuppressWarnings("ConstantConditions")
-            private String resolveSemverVersion(ExecutionContext ctx, String groupId, String artifactId) throws MavenDownloadingException {
-                if (versionComparator == null) {
-                    return newVersion;
-                }
-                if (availableVersions == null) {
-                    availableVersions = new ArrayList<>();
-                    MavenMetadata mavenMetadata = metadataFailures.insertRows(ctx, () -> downloadMetadata(groupId, artifactId, ctx));
-                    for (String v : mavenMetadata.getVersioning().getVersions()) {
-                        if (versionComparator.isValid(newVersion, v)) {
-                            availableVersions.add(v);
-                        }
-                    }
-
-                }
-                return availableVersions.isEmpty() ? newVersion : Collections.max(availableVersions, versionComparator);
             }
         };
     }
