+++ rewrite-maven/src/main/java/org/openrewrite/maven/ChangePluginGroupIdAndArtifactId.java
@@ -18,45 +18,16 @@ package org.openrewrite.maven;
 import lombok.EqualsAndHashCode;
 import lombok.Value;
 import org.openrewrite.ExecutionContext;
-import org.openrewrite.Option;
-import org.openrewrite.Recipe;
 import org.openrewrite.TreeVisitor;
-import org.openrewrite.internal.lang.Nullable;
-import org.openrewrite.maven.table.MavenMetadataFailures;
-import org.openrewrite.xml.ChangeTagValueVisitor;
 import org.openrewrite.xml.tree.Xml;
-
-import java.util.Optional;
 
 @Value
 @EqualsAndHashCode(callSuper = false)
-public class ChangePluginGroupIdAndArtifactId extends Recipe {
-    @EqualsAndHashCode.Exclude
-    MavenMetadataFailures metadataFailures = new MavenMetadataFailures(this);
+public class ChangePluginGroupIdAndArtifactId extends AbstractChangeGroupIdAndArtifactId {
 
-    @Option(displayName = "Old group ID",
-            description = "The old group ID to replace. The group ID is the first part of a plugin coordinate 'com.google.guava:guava:VERSION'. Supports glob expressions.",
-            example = "org.openrewrite.recipe")
-    String oldGroupId;
-
-    @Option(displayName = "Old artifact ID",
-            description = "The old artifactId to replace. The artifact ID is the second part of a plugin coordinate 'com.google.guava:guava:VERSION'. Supports glob expressions.",
-            example = "my-deprecated-maven-plugin")
-    String oldArtifactId;
-
-    @Option(displayName = "New group ID",
-            description = "The new group ID to use. Defaults to the existing group ID.",
-            example = "corp.internal.openrewrite.recipe",
-            required = false)
-    @Nullable
-    String newGroupId;
-
-    @Option(displayName = "New artifact ID",
-            description = "The new artifact ID to use. Defaults to the existing artifact ID.",
-            example = "my-new-maven-plugin",
-            required = false)
-    @Nullable
-    String newArtifact;
+    public ChangePluginGroupIdAndArtifactId(String oldGroupId, String oldArtifactId, String newGroupId, String newArtifactId) {
+        super(oldGroupId, oldArtifactId, newGroupId, newArtifactId);
+    }
 
     @Override
     public String getDisplayName() {
@@ -65,7 +36,7 @@ public class ChangePluginGroupIdAndArtifactId extends Recipe {
 
     @Override
     public String getInstanceNameSuffix() {
-        return String.format("`%s:%s`", newGroupId, newArtifact);
+        return String.format("`%s:%s`", newGroupId, newArtifactId);
     }
 
     @Override
@@ -76,7 +47,7 @@ public class ChangePluginGroupIdAndArtifactId extends Recipe {
     @Override
     public TreeVisitor<?, ExecutionContext> getVisitor() {
 
-        return new MavenVisitor<ExecutionContext>() {
+        return new AbstractChangeGroupIdAndArtifactIdVisitor() {
             @Override
             public Xml visitTag(Xml.Tag tag, ExecutionContext ctx) {
                 Xml.Tag t = (Xml.Tag) super.visitTag(tag, ctx);
@@ -84,8 +55,8 @@ public class ChangePluginGroupIdAndArtifactId extends Recipe {
                     if (newGroupId != null) {
                         t = changeChildTagValue(t, "groupId", newGroupId, ctx);
                     }
-                    if (newArtifact != null) {
-                        t = changeChildTagValue(t, "artifactId", newArtifact, ctx);
+                    if (newArtifactId != null) {
+                        t = changeChildTagValue(t, "artifactId", newArtifactId, ctx);
                     }
                     if (t != tag) {
                         maybeUpdateModel();
@@ -93,14 +64,6 @@ public class ChangePluginGroupIdAndArtifactId extends Recipe {
                 }
                 //noinspection ConstantConditions
                 return t;
-            }
-
-            private Xml.Tag changeChildTagValue(Xml.Tag tag, String childTagName, String newValue, ExecutionContext ctx) {
-                Optional<Xml.Tag> childTag = tag.getChild(childTagName);
-                if (childTag.isPresent() && !newValue.equals(childTag.get().getValue().orElse(null))) {
-                    tag = (Xml.Tag) new ChangeTagValueVisitor<>(childTag.get(), newValue).visitNonNull(tag, ctx);
-                }
-                return tag;
             }
         };
     }
